feat : Quantity를 추가하여 수량 포장

+++ src/main/java/kitchenpos/dto/response/MenuProductViewResponse.java
@@ -14,7 +14,7 @@ public class MenuProductViewResponse {
                 menuProduct.getSeq(),
                 menuProduct.getMenu().getId(),
                 menuProduct.getProduct().getId(),
-                menuProduct.getQuantity()
+                menuProduct.getQuantity().toLong()
         );
     }
